task2/airflow/requirements.txt
apache-airflow-providers-postgres==5.6.0
clickhouse-driver==0.2.6
psycopg2-binary==2.9.9
